Add CLI to add/edit posts.
Closes #10.

I figured that making a CLI would be so much simpler and more
straightforward than making a web UI and then having to secure it.  I
still like this approach but it's less of a slam dunk than I thought.  A
lot of stuff ended up more verbose than I thought and I felt the lack of
e.g. hidden fields, or dropdowns with different values vs. labels.